+++ CHANGELOG.rst
@@ -43,7 +43,7 @@ Bug fixes and small improvements:
 - Fix runtime problem introduced with 8.4. (:issue:`1194`)
 - Fix wrong entries in data source attribute of JSON report. (:issue:`1194`)
 - Fix nested HTML report without coverage data. (:issue:`1197`)
-- Raise error if all coverage data is filtered out. (:issue:`1200`)
+- Add warning if coverage data is empty. (:issue:`1200`)
 
 Documentation: